Safari-611.1.10/LayoutTests/js/instance-property-setter-other-instance-expected.txt

Tests that calling an instance property setter on another instance works as expected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS locationSetter is an instance of function Function() {
    [native code]
}
PASS frames[0].document.location.toString() is "about:blank"
locationSetter.call(frames[0].document, 'resources/pass-and-finish.html')
PASS Changed iframe document's location
PASS successfullyParsed is true

TEST COMPLETE
